Moved availability to separate namespace
