+++ Api/Services/Accommodations/Availability/AvailabilityResultsCache.cs
@@ -7,7 +7,7 @@ using HappyTravel.Edo.Api.Models.Markups;
 using HappyTravel.Edo.Common.Enums;
 using HappyTravel.EdoContracts.Accommodations;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations
+namespace HappyTravel.Edo.Api.Services.Accommodations.Availability
 {
     public class AvailabilityResultsCache : IAvailabilityResultsCache
     {
