+++ Api/Services/Accommodations/Availability/IAvailabilityResultsCache.cs
@@ -4,7 +4,7 @@ using HappyTravel.Edo.Api.Models.Markups;
 using HappyTravel.Edo.Common.Enums;
 using HappyTravel.EdoContracts.Accommodations;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations
+namespace HappyTravel.Edo.Api.Services.Accommodations.Availability
 {
     public interface IAvailabilityResultsCache
     {
